fix(schema): accept fractional bbox/centroid floats from B1
B1's main.py publishes bbox + centroid pixel coords rounded to 2
decimals (`round(x, 2)` in detector.py:108-117). Our Pydantic
schema declared them as `int` so every real B1 event was being
silently dropped by validate_event with 'int_from_float — got a
number with a fractional part'.

Switch BBox/Centroid to `float`, cast to `int` in the writer
(traffic_events.bbox_* columns are still INTEGER — pixel precision
below 1px isn't useful for analytics).

Locked in by a new test_b1_float_bbox_centroid in test_validator.py
that feeds the exact shape main.py emits.

Confirmed by manual smoke test against the validator with the
B1-style payload from services/b1-detection/src/main.py:127-138.

+++ src/shared/schemas.py
@@ -6,15 +6,16 @@ from pydantic import BaseModel, Field
 # --- Input event from B1 via Kafka ---
 
 class BBox(BaseModel):
-    x: int
-    y: int
-    w: int
-    h: int
+    # B1 rounds pixel coords to 2 decimals; accept floats and cast in the writer.
+    x: float
+    y: float
+    w: float
+    h: float
 
 
 class Centroid(BaseModel):
-    x: int
-    y: int
+    x: float
+    y: float
 
 
 class TrafficEventInput(BaseModel):
